Shark Attack Analytics M vs F - Python 3
This is a for fun project I did using a free shark attack dataset because I had an argument with a friend that a female has to be more likely to be attacked due to their menstrual cycle and average higher fat content.... My analysis proved me wrong. I thought it would be astoundingly greater.